+++ crates/krb5-client/src/bin/krb5-kvno.rs
@@ -177,6 +177,7 @@ fn run(
         cname: cred.client.1.clone(),
         crealm: cred.client.0.clone(),
         fast_avail: false,
+        used_fast: false,
         pa_type: None,
     };
     let tgs = if let Some(who) = args.for_user.as_deref() {